+++ SharpAlliance.Core/Morale.cs
@@ -482,7 +482,11 @@ public class Morale
                         // CJC: adding to SOLDIER_IN_SECTOR check special stuff because the old sector values might
                         // be appropriate (because in transit going out of that sector!)
 
-                        if (SOLDIER_IN_SECTOR(pTeamSoldier, sMapX, sMapY, bMapZ) || (pTeamSoldier.fBetweenSectors && ((pTeamSoldier.ubPrevSectorID % 16) + 1) == sMapX && ((MAP_ROW)(pTeamSoldier.ubPrevSectorID / 16) + 1) == sMapY && (pTeamSoldier.bSectorZ == bMapZ)))
+                        if (SOLDIER_IN_SECTOR(pTeamSoldier, sMapX, sMapY, bMapZ)
+                            || (pTeamSoldier.fBetweenSectors
+                                && (((int)pTeamSoldier.ubPrevSectorID % 16) + 1) == sMapX
+                                && ((MAP_ROW)((int)pTeamSoldier.ubPrevSectorID / 16) + 1) == sMapY
+                                && (pTeamSoldier.bSectorZ == bMapZ)))
                         {
                             switch (gMercProfiles[pTeamSoldier.ubProfile].bAttitude)
                             {
